refactor(protocol): hoist the shared `Wire*` structs into `city-sim-protocol`
`WireBuilding`, `WireUtilityComponent`, `WireEducationStats`,
`WireEducationSeatsUsed`, and `WireBudgetHistoryEntry` existed as
byte-identical copies in `city-sim-wasm` and `tauri-plugin-city-sim` —
two hand-synchronised definitions of what is one wire contract, feeding
two separately hand-written TS mirrors. One canonical copy now lives in
`city_sim_protocol::wire_types` and both hosts import it. Pure move: no
field, type, or serialisation change. The `From` conversions from the
engine's own types live in `city-sim-core`'s `wire.rs` rather than next
to the structs — `city-sim-protocol` does not depend on
`city-sim-core`, and Rust's orphan rules require the impls to sit where
the source types are local.
